[ADD] collectionview 하나의 view에 통합
