+++ Maddori.Apple/Maddori.Apple/Screen/Home/SelectReflectionMember/View/SelectReflectionMemberView.swift
@@ -12,6 +12,32 @@ import RxCocoa
 
 final class SelectReflectionMemberView: UIView {
     
+    private enum Size {
+        static let collectionHorizontalSpacing: CGFloat = 34
+        static let collectionVerticalSpacing: CGFloat = 20
+        static let cellWidth: CGFloat = UIScreen.main.bounds.width * 0.37
+        static let cellHeight: CGFloat = 140
+        static let cellInteritemSpacingSpacing: CGFloat = 27
+        static let cellLineSpacing: CGFloat = 20
+        static let collectionInsets = UIEdgeInsets(
+            top: collectionVerticalSpacing,
+            left: collectionHorizontalSpacing,
+            bottom: collectionVerticalSpacing,
+            right: collectionHorizontalSpacing)
+    }
+    
+    var memberList: [MemberDetailResponse] = [] {
+        didSet {
+            memberCollectionView.reloadData()
+        }
+    }
+    var didTappedMember: ((MemberDetailResponse, [Int]) -> ())?
+    var selectedMemberList: [Int] = UserDefaultStorage.seenMemberIdList {
+        willSet {
+            UserDefaultHandler.appendSeenMemberIdList(memberIdList: newValue)
+        }
+    }
+    
     // MARK: - ui components
     
     private let closeButton = CloseButton()
@@ -28,7 +54,25 @@ final class SelectReflectionMemberView: UIView {
         label.textColor = .gray400
         return label
     }()
-    let memberCollectionView = ReflectionMemberCollectionView()
+    private let collectionViewFlowLayout: UICollectionViewFlowLayout = {
+        let flowLayout = UICollectionViewFlowLayout()
+        flowLayout.scrollDirection = .vertical
+        flowLayout.sectionInset = Size.collectionInsets
+        flowLayout.itemSize = CGSize(width: Size.cellWidth, height: Size.cellHeight)
+        flowLayout.minimumInteritemSpacing = Size.cellInteritemSpacingSpacing
+        flowLayout.minimumLineSpacing = Size.cellLineSpacing
+        return flowLayout
+    }()
+    lazy var memberCollectionView: UICollectionView = {
+        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: collectionViewFlowLayout)
+        collectionView.dataSource = self
+        collectionView.delegate = self
+        collectionView.backgroundColor = .white200
+        collectionView.showsVerticalScrollIndicator = false
+        collectionView.allowsMultipleSelection = true
+        collectionView.register(ReflectionMemberCollectionViewCell.self, forCellWithReuseIdentifier: ReflectionMemberCollectionViewCell.className)
+        return collectionView
+    }()
     let feedbackDoneButton: MainButton = {
         let button = MainButton()
         button.isDisabled = true
@@ -71,12 +115,12 @@ final class SelectReflectionMemberView: UIView {
     }
     
     func setTeamMembers(teamMembers: [MemberDetailResponse]) {
-        memberCollectionView.memberList = teamMembers
-        feedbackDoneButton.title = TextLiteral.selectReflectionMemberViewControllerDoneButtonText + "(\(UserDefaultStorage.seenMemberIdList.count)/\(memberCollectionView.memberList.count))"
+        memberList = teamMembers
+        feedbackDoneButton.title = TextLiteral.selectReflectionMemberViewControllerDoneButtonText + "(\(UserDefaultStorage.seenMemberIdList.count)/\(memberList.count))"
     }
     
     func setupPreviousStatus() {
-        feedbackDoneButton.title = TextLiteral.selectReflectionMemberViewControllerDoneButtonText + "(\(UserDefaultStorage.seenMemberIdList.count)/\(memberCollectionView.memberList.count))"
+        feedbackDoneButton.title = TextLiteral.selectReflectionMemberViewControllerDoneButtonText + "(\(UserDefaultStorage.seenMemberIdList.count)/\(memberList.count))"
         feedbackDoneButton.isDisabled = !UserDefaultStorage.completedCurrentReflection
     }
 }
@@ -110,5 +154,46 @@ extension SelectReflectionMemberView {
             $0.bottom.equalTo(feedbackDoneButton.snp.top).inset(-6)
             $0.leading.trailing.equalToSuperview()
         }
+    }
+}
+
+// MARK: - collectionView extension
+
+extension SelectReflectionMemberView: UICollectionViewDelegate {
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        let selectedMember = memberList[indexPath.item]
+        if !selectedMemberList.contains(where: { $0 == selectedMember.id }) {
+            selectedMemberList.append(selectedMember.id ?? 0)
+        }
+        didTappedMember?(selectedMember, selectedMemberList)
+    }
+}
+
+extension SelectReflectionMemberView: UICollectionViewDataSource {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return memberList.count
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: ReflectionMemberCollectionViewCell.className, for: indexPath) as? ReflectionMemberCollectionViewCell else {
+            assert(false, "Wrong Cell")
+            return UICollectionViewCell()
+        }
+        
+        if let imagePath = memberList[indexPath.item].profileImagePath {
+            let fullImagePath = UrlLiteral.imageBaseURL + imagePath
+            cell.profileImage.load(from: fullImagePath)
+        }
+        
+        cell.nicknameLabel.text = memberList[indexPath.item].nickname
+        cell.roleLabel.text = memberList[indexPath.item].role
+                
+        if let userId = memberList[indexPath.item].id {
+            if selectedMemberList.contains(userId) {
+                cell.applySelectedAttribute()
+            }
+        }
+        
+        return cell
     }
 }
